Apex (#6)
* register func

* err msg

+++ model.py
@@ -132,7 +132,7 @@ class GWNet(nn.Module):
         adjacency_matrices = self.fixed_supports
         # calculate the current adaptive adj matrix once per iteration
         if self.addaptadj:
-            adp = F.softmax(F.relu(torch.mm(self.nodevec1, self.nodevec2)) / self.softmax_temp, dim=1)
+            adp = F.softmax(F.relu(torch.mm(self.nodevec1, self.nodevec2)) / self.softmax_temp, dim=1).half()
             adjacency_matrices = self.fixed_supports + [adp]
 
         # WaveNet layers
